Adopt socket logic to Windows
Split the socket logic into unix and windows files and have the windows
sockets use named pipes instead. Unix sockets should work on Windows in
theory but they caused a lot of problems with dangling sockets and
permission problems. The named pipes however, work just fine.
