+++ pkg/client/connector/command.go
@@ -2,14 +2,11 @@ package connector
 
 import (
 	"context"
-	"errors"
-	"fmt"
 	"net"
 	"os"
 	"path/filepath"
 	"sort"
 	"sync"
-	"syscall"
 	"time"
 
 	"github.com/spf13/cobra"
@@ -355,17 +352,10 @@ func run(c context.Context) error {
 		}()
 
 		// Listen on unix domain socket
-		grpcListener, err = net.Listen("unix", client.ConnectorSocketName)
+		grpcListener, err = client.ListenSocket(c, ProcessName, client.ConnectorSocketName)
 		if err != nil {
-			if errors.Is(err, syscall.EADDRINUSE) {
-				return fmt.Errorf("socket %q exists so the %s is either already running or terminated ungracefully",
-					client.SocketURL(client.ConnectorSocketName), ProcessName)
-			}
 			return err
 		}
-		// Don't have dhttp.ServerConfig.Serve unlink the socket; defer unlinking the socket
-		// until the process exits.
-		grpcListener.(*net.UnixListener).SetUnlinkOnClose(false)
 
 		dlog.Info(c, "gRPC server started")
 		defer func() {
