+++ pkg/client/daemon/service.go
@@ -8,12 +8,9 @@ import (
 	"net/http"
 	"os"
 	"path/filepath"
-	"syscall"
 	"time"
 
-	"github.com/pkg/errors"
 	"github.com/spf13/cobra"
-	"golang.org/x/sys/unix"
 	"google.golang.org/grpc"
 	empty "google.golang.org/protobuf/types/known/emptypb"
 
@@ -27,6 +24,7 @@ import (
 	"github.com/telepresenceio/telepresence/v2/pkg/client"
 	"github.com/telepresenceio/telepresence/v2/pkg/client/logging"
 	"github.com/telepresenceio/telepresence/v2/pkg/filelocation"
+	"github.com/telepresenceio/telepresence/v2/pkg/proc"
 )
 
 const ProcessName = "daemon"
@@ -97,8 +95,8 @@ func (d *service) SetOutboundInfo(ctx context.Context, info *rpc.OutboundInfo) (
 
 // run is the main function when executing as the daemon
 func run(c context.Context, loggingDir, configDir, dns string) error {
-	if os.Geteuid() != 0 {
-		return fmt.Errorf("telepresence %s must run as root", ProcessName)
+	if !proc.IsAdmin() {
+		return fmt.Errorf("telepresence %s must run with elevated privileges", ProcessName)
 	}
 
 	// Spoof the AppUserLogDir and AppUserConfigDir so that they return the original user's
@@ -190,19 +188,10 @@ func run(c context.Context, loggingDir, configDir, dns string) error {
 
 		// Listen on unix domain socket
 		dlog.Debug(c, "gRPC server starting")
-		origUmask := unix.Umask(0)
-		grpcListener, err = net.Listen("unix", client.DaemonSocketName)
-		unix.Umask(origUmask)
+		grpcListener, err = client.ListenSocket(c, ProcessName, client.DaemonSocketName)
 		if err != nil {
-			if errors.Is(err, syscall.EADDRINUSE) {
-				return fmt.Errorf("socket %q exists so the %s is either already running or terminated ungracefully",
-					client.SocketURL(client.DaemonSocketName), ProcessName)
-			}
 			return err
 		}
-		// Don't have dhttp.ServerConfig.Serve unlink the socket; defer unlinking the socket
-		// until the process exits.
-		grpcListener.(*net.UnixListener).SetUnlinkOnClose(false)
 
 		dlog.Info(c, "gRPC server started")
 		defer func() {
