+++ pkg/client/sockets.go
@@ -1,30 +1,9 @@
 package client
 
 import (
-	"context"
-	"errors"
 	"fmt"
-	"net"
-	"os"
-	"syscall"
 	"time"
-
-	"google.golang.org/grpc"
 )
-
-const (
-	// ConnectorSocketName is the path used when communicating to the connector process
-	ConnectorSocketName = "/tmp/telepresence-connector.socket"
-
-	// DaemonSocketName is the path used when communicating to the daemon process
-	DaemonSocketName = "/var/run/telepresence-daemon.socket"
-)
-
-// SocketExists returns true if a socket is found at the given path
-func SocketExists(path string) bool {
-	s, err := os.Stat(path)
-	return err == nil && s.Mode()&os.ModeSocket != 0
-}
 
 // WaitUntilSocketVanishes waits until the socket at the given path is removed
 // and returns when that happens. The wait will be max ttw (time to wait) long.
@@ -32,12 +11,8 @@ func SocketExists(path string) bool {
 func WaitUntilSocketVanishes(name, path string, ttw time.Duration) (err error) {
 	giveUp := time.Now().Add(ttw)
 	for giveUp.After(time.Now()) {
-		_, err = os.Stat(path)
-		if err != nil {
-			if os.IsNotExist(err) {
-				err = nil
-			}
-			return err
+		if !SocketExists(path) {
+			return nil
 		}
 		time.Sleep(250 * time.Millisecond)
 	}
@@ -50,58 +25,10 @@ func WaitUntilSocketVanishes(name, path string, ttw time.Duration) (err error) {
 func WaitUntilSocketAppears(name, path string, ttw time.Duration) (err error) {
 	giveUp := time.Now().Add(ttw)
 	for giveUp.After(time.Now()) {
-		_, err = os.Stat(path)
-		if err == nil {
-			return
-		}
-		if !os.IsNotExist(err) {
-			return err
+		if SocketExists(path) {
+			return nil
 		}
 		time.Sleep(250 * time.Millisecond)
 	}
 	return fmt.Errorf("timeout while waiting for %s to start", name)
-}
-
-// SocketURL returns the URL that corresponds to the given unix socket filesystem path.
-func SocketURL(socket string) string {
-	// The unix URL scheme was implemented in google.golang.org/grpc v1.34.0
-	return "unix:" + socket
-}
-
-// DialSocket dials the given unix socket and returns the resulting connection
-func DialSocket(ctx context.Context, socketName string, opts ...grpc.DialOption) (*grpc.ClientConn, error) {
-	ctx, cancel := context.WithTimeout(ctx, 5*time.Second) // FIXME(lukeshu): Make this configurable
-	defer cancel()
-	conn, err := grpc.DialContext(ctx, SocketURL(socketName), append([]grpc.DialOption{
-		grpc.WithInsecure(),
-		grpc.WithNoProxy(),
-		grpc.WithBlock(),
-		grpc.FailOnNonTempDialError(true),
-	}, opts...)...)
-	if err != nil {
-		if err == context.DeadlineExceeded {
-			// grpc.DialContext doesn't wrap context.DeadlineExceeded with any useful
-			// information at all.  Fix that.
-			err = &net.OpError{
-				Op:  "dial",
-				Net: "unix",
-				Addr: &net.UnixAddr{
-					Name: socketName,
-					Net:  "unix",
-				},
-				Err: fmt.Errorf("socket exists but is not responding: %w", err),
-			}
-		}
-		// Add some Telepresence-specific commentary on what specific common errors mean.
-		switch {
-		case errors.Is(err, context.DeadlineExceeded):
-			err = fmt.Errorf("%w; this usually means that the process has locked up", err)
-		case errors.Is(err, syscall.ECONNREFUSED):
-			err = fmt.Errorf("%w; this usually means that the process has terminated ungracefully", err)
-		case errors.Is(err, os.ErrNotExist):
-			err = fmt.Errorf("%w; this usually means that the process is not running", err)
-		}
-		return nil, err
-	}
-	return conn, nil
 }
